+++ cmd/root.go
@@ -40,6 +40,9 @@ const (
 
 func init() {
 	rootCmd.PersistentFlags().StringP("output", "o", ".", "Output directory path")
+	rootCmd.Flags().StringP("chart-repo-prefix", "", "charts/giantswarm", "Prefix for chart repositories in the OCI registries")
+	rootCmd.Flags().StringP("public-oci-registry", "", "gsoci.azurecr.io", "Host name of the public OCI registry")
+	rootCmd.Flags().StringP("private-oci-registry", "", "gsociprivate.azurecr.io", "Host name of the private OCI registry")
 
 	rootCmd.AddCommand(appcatalogs.Command)
 	rootCmd.AddCommand(installations.Command)
@@ -55,6 +58,24 @@ func Execute() {
 
 func runRoot(cmd *cobra.Command, args []string) {
 	path, err := cmd.PersistentFlags().GetString("output")
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	repoPrefix, err := cmd.Flags().GetString("chart-repo-prefix")
+	if err != nil {
+		log.Fatal(err)
+	}
+	// Remove leading and trailing slash if present
+	repoPrefix = strings.TrimPrefix(repoPrefix, "/")
+	repoPrefix = strings.TrimSuffix(repoPrefix, "/")
+
+	publicOciRegistry, err := cmd.Flags().GetString("public-oci-registry")
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	privateOciRegistry, err := cmd.Flags().GetString("private-oci-registry")
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -102,9 +123,14 @@ func runRoot(cmd *cobra.Command, args []string) {
 		log.Printf("Processing %d repos of team %q\n", len(list.Repositories), list.OwnerTeamName)
 
 		for _, repo := range list.Repositories {
+
+			ociRegistry := publicOciRegistry
 			isPrivate, err := repoService.GetIsPrivate(repo.Name)
 			if err != nil {
 				log.Fatalf("Error: %v", err)
+			}
+			if isPrivate {
+				ociRegistry = privateOciRegistry
 			}
 
 			hasReadme, err := repoService.GetHasReadme(repo.Name)
@@ -114,6 +140,7 @@ func runRoot(cmd *cobra.Command, args []string) {
 
 			// Fetch Helm chart info if available.
 			var charts []*helmchart.Chart
+			var hasDeployableChart bool
 			{
 				numCharts, err := repoService.GetNumHelmCharts(repo.Name)
 				if err != nil {
@@ -134,6 +161,10 @@ func runRoot(cmd *cobra.Command, args []string) {
 								log.Printf("WARN - %s - error parsing helm chart %s: %v", repo.Name, chartName, err)
 							} else {
 								charts = append(charts, chart)
+							}
+
+							if chart.Type == "application" || chart.Type == "" {
+								hasDeployableChart = true
 							}
 						}
 					}
@@ -183,18 +214,21 @@ func runRoot(cmd *cobra.Command, args []string) {
 				genLanguage = string(repo.Gen.Language)
 			}
 
-			genFlavors := []string{}
+			genFlavors := make([]string, len(repo.Gen.Flavors))
 			for i, flavor := range repo.Gen.Flavors {
 				genFlavors[i] = string(flavor)
 			}
 
 			c, err := component.New(
 				repo.Name,
+				component.WithCircleCiSlug(fmt.Sprintf("github/%s/%s", githubOrganization, repo.Name)),
 				component.WithDefaultBranch(defaultBranch),
 				component.WithDependsOn(deps...),
 				component.WithDeploymentNames(deploymentNames...),
 				component.WithDescription(description),
 				component.WithFlavors(genFlavors...),
+				component.WithGithubProjectSlug(fmt.Sprintf("%s/%s", githubOrganization, repo.Name)),
+				component.WithGithubTeamSlug(list.OwnerTeamName),
 				component.WithHasReadme(hasReadme),
 				component.WithHasReleases(latestReleaseTag != ""),
 				component.WithHelmCharts(charts...),
@@ -206,9 +240,15 @@ func runRoot(cmd *cobra.Command, args []string) {
 				component.WithPrivate(isPrivate),
 				component.WithSystem(repo.System),
 				component.WithType(repo.ComponentType),
+				component.WithOciRegistry(ociRegistry),
+				component.WithOciRepositoryPrefix(repoPrefix),
 			)
 			if err != nil {
 				log.Fatalf("Could not create component: %s", err)
+			}
+
+			if hasDeployableChart {
+				c.AddTag("helmchart-deployable")
 			}
 
 			// Grafana dashboard link for services.